T.1 + T.12–T.19: multidtype value assertions + remaining tests/tooling
audit-3 final wave.

T.1 (P0, ~60 multidtype tests previously asserting only shape) —
landed in 3 parallel batches:

Batch 1 (comparison / indexing / losses / linalg / pool3d / vit):
* test_comparison_operators_multidtype — value-level expected for
  EqualSelf / NotEqual / LessThan / GreaterThan / LessEqual /
  GreaterEqual / BroadcastComparison via new
  expectBoolTensorEquals helper.
* test_indexing_operator_multidtype — SelectDim0/1, NarrowDim0,
  SliceBasic, WhereCondition, MaskedFill now CPU-reference or
  inline closed-form.
* test_losses_missing_multidtype — merged with T.8 (numel<=1 →
  numel==1) and added scalar value match vs CPU+F32 reference
  across 18 TEST_Ps covering MarginRanking / PoissonNLL /
  CosineEmbedding / TripletMargin / GaussianNLL / SoftMargin /
  HingeEmbedding / MultiLabelSoftMargin / MultiMargin.
* test_linalg_missing_multidtype — 27 TEST_Ps now assert numerics
  via CPU reference or closed-form identity (det / slogdet / inv
  → A·A⁻¹=I / cholesky reconstruction / QR → Q·R=A / SVD →
  UΣVᴴ=A + non-neg singular values / eigh → Vdiag(λ)Vᵀ=A /
  LU + lu_solve round-trip / solve / triangular_solve /
  cholesky_solve / ldl_factor_solve / pinv Moore-Penrose
  property 1 / vector_norm / vec_dot / geqrf / ormqr /
  householder_product).
* test_pooling3d_multidtype — MaxPool3d / AvgPool3d /
  AdaptiveMaxPool3d / AdaptiveAvgPool3d forward + gradient now
  CPU-reference; AdaptiveAvgPool3dGlobalPooling uses closed-form
  spatial-mean reference.
* test_vit_multidtype — new expectOutputSane helper applied to
  19 forward-shape TEST_Ps across ViT-Base/Large/Huge variants
  plus train/eval mode + batch / channel / image-size variations.

Batch 2 (ZeRO + nn-optim):
* test_zero_stage1/2/3_multidtype + their integration variants +
  test_zero_profiling_multidtype + test_zero_stage1_distributed_
  multidtype — every shape-only TEST_P now asserts param
  trajectory vs CPU+F32 plain-Adam reference (Step,
  MultipleStepsTrajectory, StateDictRoundtrip, etc.) or
  closed-form Adam-with-known-gradient value.
* test_lbfgs_multidtype — already had convergence assertions;
  added Rosenbrock-optimum location check + file-level rationale
  comment.

Batch 3 (autograd / bert / electra / models + infra):
* New shared fixture helper expectFiniteNonZero
  (multi_backend_dtype_fixture.hpp) — used in model-level tests
  where full CPU reference is prohibitive but NaN/Inf/all-zero
  failure modes need surfacing.
* test_bmm_autograd / test_conv1d / test_convtranspose2d /
  test_kldiv_manual / test_mfcc / test_rnn_cells / test_roi_ops /
  test_split_operation — CPU-reference forward and gradient
  cross-checks on every shape-only TEST_P.
* test_albert_t5, test_bert, test_electra,
  test_attention_sdpa, test_classic_models, test_deeplabv3plus,
  test_detection_components, test_detection_ops,
  test_float16_regression, test_sparse_autograd — value-level
  finite-nonzero on previously shape-only model-forward tests.
* test_autocast and test_comparison_ops files received rationale
  comments — every TEST_F there already asserts on element-level
  state-machine return values; preserved audit-2 P.9 pattern.

T.12 — test_phase11_backends now `BACKENDS multi DISCOVERY_MODE
PRE_TEST`.

T.13 — test_linalg_lstsq_pinv_matexp_multidtype gained
PinvIdentity_Complex64 and PinvRankDeficient (Moore-Penrose
property 1 verified on 3×3 rank-2 input).

T.14 — tools/check_multidtype_coverage.py + ctest wiring
(`LABELS lint`).  MULTIDTYPE_COVERAGE_GAPS.md table converted to
a machine-readable block delimited by
KNOWN-INTENTIONAL-START/END markers; 21 aspirational + 82
grandfathered pre-existing gaps included so the ratchet currently
exits 0 and trips on new gaps.

T.15 — test_higher_order_activations_multidtype 5
EXPECT_NO_THROW(loss.backward(...)) lines (create_graph variant)
replaced with proper `loss.backward(...); EXPECT_GRAD_FLOWS(...)`
pairs.  Macro surfaces real ROCm-only severed-grad bugs in
Hardswish / Hardsigmoid / LeakyReLU paths (queued as follow-ups).

T.16 — TESTING.md:291-295 rewritten to reflect actual behaviour:
--gtest_also_run_disabled_tests is applied to every target;
DISABLED_ in this project means "performance-ceiling regression
assertion test"; real skipping requires GTEST_SKIP().

T.17 — new third_party/THIRD_PARTY.md pinning nlohmann/json
v3.11.3 (pointer to in-file comment), iree_dist + iree_compat
vendored snapshot dates + MLIR-LLVM upgrade-gate note re:
src/jit/codegen.cpp, transitive spdlog / fmt / pybind11
versions.

T.18 — tools/op_coverage_baseline.json regenerated; literal
"unknown" entries stripped; counts updated (CPU 317 → 468 ops —
the baseline was staler than the audit doc reported).

T.19 — src/backends/cpu/kernels/nn_kernels.cpp:50 replaced the
static thread_local tl_rng with an accessor that reseeds via
tenzor::get_global_seed() whenever the global manual seed flips.
TESTING.md "Reproducibility" section documents the canonical seed
entry point (`tenzor::manual_seed(uint64_t)`).

Coverage notes: T.1 batch 3 left ~15 model files
(test_efficientnet / test_fused_ops / test_fusion_optimizer /
test_gpt / test_graph_optimizer / test_gru / test_lstm /
test_mask_rcnn / test_mobilenet_v2_v3 / test_rnn /
test_swin_transformer / test_transforms / test_unet /
test_vision_components / test_yolo plus tests/autograd/
*_multidtype.cpp) untouched — pattern is mechanical (apply
expectFiniteNonZero to forward outputs) and will land in a
follow-up commit.  T.15 found real ROCm Hardswish / Hardsigmoid /
LeakyReLU severed-grad bugs needing their own fix.

Verified: full ninja build clean in single -j1 shell.

+++ src/jit/graph.cpp
@@ -1832,8 +1832,8 @@ auto Graph::execute_node(const std::shared_ptr<Node>& node,
 
         case OpType::Clamp:
             if (!input_vars.empty()) {
-                float min_val = node->get_attr("min");
-                float max_val = node->get_attr("max");
+                double min_val = static_cast<double>(static_cast<float>(node->get_attr("min")));
+                double max_val = static_cast<double>(static_cast<float>(node->get_attr("max")));
                 outputs.push_back(tenzor::clamp(input_vars[0], min_val, max_val));
             }
             break;
